+++ source/_internals/testUtils.js
@@ -1,5 +1,14 @@
 import combinate from 'combinate'
-import { equals, filter, forEach, omit, switcher, type } from 'rambdax'
+import {
+  equals,
+  filter,
+  forEach,
+  init,
+  last,
+  omit,
+  switcher,
+  type,
+} from 'rambdax'
 
 const omitOk = omit('ok')
 
@@ -39,11 +48,15 @@ function parseError(err){
   }
 }
 
-function executeSync(fn, inputs){
+function executeSync(
+  fn, inputs, returnsFunctionFlag
+){
   let result = PENDING
   let error = { ok : false }
   try {
-    result = fn(...inputs)
+    result = returnsFunctionFlag ?
+      fn(...init(inputs))(last(inputs)) :
+      fn(...inputs)
   } catch (e){
     error = parseError(e)
   }
@@ -134,11 +147,16 @@ export function profileMethodAsync({
   })
 }
 
-export function compareToRamda(fn, fnRamda){
+export function compareToRamda(
+  fn, fnRamda, returnsFunctionFlag
+){
   return (...inputs) => {
     const { result, error } = executeSync(fn, inputs)
-    const { result: ramdaResult, error: ramdaError } = executeSync(fnRamda,
-      inputs)
+    const { result: ramdaResult, error: ramdaError } = executeSync(
+      fnRamda,
+      inputs,
+      returnsFunctionFlag
+    )
 
     const toReturn = {
       result,
@@ -204,6 +222,7 @@ export const compareCombinations = ({
   firstInput,
   secondInput = undefined,
   thirdInput = undefined,
+  returnsFunctionFlag = false,
   setCounter = () => {},
   callback = x => {},
   fn,
@@ -233,7 +252,9 @@ export const compareCombinations = ({
   })
   const inputKeys = Object.keys(combinationsInput)
   const combinations = combinate(combinationsInput)
-  const compareOutputs = compareToRamda(fn, fnRamda)
+  const compareOutputs = compareToRamda(
+    fn, fnRamda, returnsFunctionFlag
+  )
 
   afterAll(() => callback(counter))