+++ source/tryCatch.spec.js
@@ -1,134 +1,118 @@
-import { delay } from './delay'
-import { prop } from './prop'
-import { tryCatch } from './tryCatch'
+import { compareCombinations } from "./_internals/testUtils";
+import { prop } from "./prop";
+import { tryCatch } from "./tryCatch";
+import { tryCatch as tryCatchRamda } from "ramda";
 
-test('throws when fn is not function', () => {
-  const fn = 'foo'
+test("throws when fn is not function", () => {
+  const fn = "foo";
 
-  expect(() => tryCatch(fn, false)(null)).toThrowWithMessage(Error,
-    'R.tryCatch | fn \'foo\'')
-})
+  expect(() => tryCatch(fn, false)(null)).toThrowWithMessage(
+    Error,
+    "R.tryCatch | fn 'foo'"
+  );
+});
 
-test('when fallback is used', () => {
-  const fn = x => x.x
+test("when fallback is used", () => {
+  const fn = (x) => x.x;
 
-  expect(tryCatch(fn, false)(null)).toBeFalse()
-})
+  expect(tryCatch(fn, false)(null)).toBeFalse();
+});
 
-test('with json parse', () => {
-  const good = () => JSON.parse(JSON.stringify({ a : 1 }))
-  const bad = () => JSON.parse('a{a')
+test("with json parse", () => {
+  const good = () => JSON.parse(JSON.stringify({ a: 1 }));
+  const bad = () => JSON.parse("a{a");
 
-  expect(tryCatch(good, 1)(null)).toEqual({ a : 1 })
-  expect(tryCatch(bad, 1)(null)).toBe(1)
-})
+  expect(tryCatch(good, 1)(null)).toEqual({ a: 1 });
+  expect(tryCatch(bad, 1)(null)).toBe(1);
+});
 
-test('when fallback is function', () => {
-  const fn = x => x.x
+test("when fallback is function", () => {
+  const fn = (x) => x.x;
 
-  expect(tryCatch(fn, () => 1)(null)).toBe(1)
-})
+  expect(tryCatch(fn, () => 1)(null)).toBe(1);
+});
 
-test('when fn is used', () => {
-  const fn = prop('x')
+test("when fn is used", () => {
+  const fn = prop("x");
 
-  expect(tryCatch(fn, false)({})).toBe(undefined)
-  expect(tryCatch(fn, false)({ x : 1 })).toBe(1)
-})
+  expect(tryCatch(fn, false)({})).toBe(undefined);
+  expect(tryCatch(fn, false)({ x: 1 })).toBe(1);
+});
 
-test('fallback receives error object and all initial inputs', () => {
-  function thrower(
-    a, b, c
-  ){
-    void c
-    throw new Error('throwerError')
+test("fallback receives error object and all initial inputs", () => {
+  function thrower(a, b, c) {
+    void c;
+    throw new Error("throwerError");
   }
 
-  function catchFn(
-    e, a, b, c
-  ){
-    return [ e.message, a, b, c ].join('|')
+  function catchFn(e, a, b, c) {
+    return [e.message, a, b, c].join("|");
   }
 
-  const willThrow = tryCatch(thrower, catchFn)
-  const result = willThrow(
-    'A', 'B', 'C'
-  )
-  expect(result).toBe('throwerError|A|B|C')
-})
+  const willThrow = tryCatch(thrower, catchFn);
+  const result = willThrow("A", "B", "C");
+  expect(result).toBe("throwerError|A|B|C");
+});
 
-test('fallback receives error object', () => {
-  function throwFn(){
-    throw new Error(10)
+test("fallback receives error object", () => {
+  function throwFn() {
+    throw new Error(10);
   }
 
-  function eCatcher(
-    e, a, b
-  ){
-    return e.message
+  function eCatcher(e, a, b) {
+    return e.message;
   }
 
-  const willThrow = tryCatch(throwFn, eCatcher)
-  expect(willThrow([])).toBe('10')
-  expect(willThrow([ {}, {}, {} ])).toBe('10')
-})
+  const willThrow = tryCatch(throwFn, eCatcher);
+  expect(willThrow([])).toBe("10");
+  expect(willThrow([{}, {}, {}])).toBe("10");
+});
 
-test('when async + fallback', async () => {
-  let called = false
+const possibleFns = [
+  null,
+  () => 1,
+  () => 0,
+  () => JSON.parse("{a:1"),
+  () => {
+    const x = {};
 
-  const fn = async input => {
-    await delay(input)
-    called = true
+    return x.x;
+  },
+  (x) => x.foo,
+  () => {
+    throw new Error("foo");
+  },
+];
 
-    return JSON.parse('{a:')
-  }
+const possibleCatchers = [
+  null,
+  (e) => e.message.length,
+  (e, ...inputs) => `${e.message.length} ${inputs.length}`,
+  () => {
+    throw new Error("bar");
+  },
+];
 
-  expect(await tryCatch(fn, 'fallback')(100)).toBe('fallback')
-  expect(called).toBeTrue()
-})
+const possibleInputs = [null, {}, { foo: 1 }];
 
-test('when async + fallback is function', async () => {
-  let called = false
-
-  const fn = async input => {
-    await delay(input)
-    called = true
-
-    return JSON.parse('{a:')
-  }
-
-  expect(await tryCatch(fn, x => x + 1)(100)).toBe(101)
-  expect(called).toBeTrue()
-})
-
-test('when async + fallback is async', async () => {
-  let called = false
-  const fn = async input => {
-    await delay(input)
-    called = true
-
-    return JSON.parse('{a:')
-  }
-  const fallback = async input => {
-    await delay(10)
-
-    return input + 1
-  }
-
-  expect(await tryCatch(fn, fallback)(100)).toBe(101)
-  expect(called).toBeTrue()
-})
-
-test('when async + fn', async () => {
-  let called = false
-
-  const fn = async input => {
-    await delay(input)
-    called = true
-
-    return input + 1
-  }
-
-  expect(await tryCatch(fn, 'fallback')(100)).toBe(101)
-  expect(called).toBeTrue()
-})
+describe("brute force", () => {
+  compareCombinations({
+    returnsFunctionFlag: true,
+    firstInput: possibleFns,
+    callback: (errorsCounters) => {
+      expect(errorsCounters).toMatchInlineSnapshot(`
+        Object {
+          "ERRORS_MESSAGE_MISMATCH": 0,
+          "ERRORS_TYPE_MISMATCH": 12,
+          "RESULTS_MISMATCH": 58,
+          "SHOULD_NOT_THROW": 0,
+          "SHOULD_THROW": 14,
+        }
+      `);
+    },
+    secondInput: possibleCatchers,
+    thirdInput: possibleInputs,
+    fn: tryCatch,
+    fnRamda: tryCatchRamda,
+  });
+});