+++ scripts/populate-readme-data/assets/CHANGELOG.md
@@ -4,6 +4,8 @@ WIP 6.4.0
 
 - Remove file extension in `main` property in `package.json` in order to allow `experimental-modules`. See also this Ramda's PR - https://github.com/ramda/ramda/pull/2678/files
 
+- Import `R.indexBy` changes from recent `@types/ramda` release.
+
 6.3.1
 
 - Fix missing `Evolved` declaration in Typescript definition