+++ scripts/populate-readme-data/assets/CHANGELOG.md
@@ -1,8 +1,10 @@
-WIP 6.4.0
+WIP 6.5.0
+
+- Close [Issue #553](https://github.com/selfrefactor/rambda/issues/553) - problem with curried typings of `R.prop`
+
+6.4.0
 
 - Upgrade all `rollup` related dependencies
-
-- Close [Issue #553](https://github.com/selfrefactor/rambda/issues/553) - ...
 
 - `R.type` supports `Symbol` just like *Ramda*.